Create README.md
In this project my objective is to transform a text into a format in which we can predict what character it is based on the dialogues of a series. For this we will use web scrapping to be able to scrape the dialogues and advanced techniques such as: Normalize, stemize, tokenize and vectorize.